# apps/m01-caching-fundamentals/openapi.json
{"openapi":"3.0.3","info":{"title":"m01-caching-fundamentals","description":"Module 1 — two-layer cache-aside (lru-cache + Redis) with HTTP caching headers","version":"1.0.0"},"components":{"schemas":{}},"paths":{"/health":{"get":{"summary":"Report redis-cache status","tags":["health"],"description":"Always 200 while the process runs: product reads fall back to the database when Redis is down, so the instance can still serve traffic. `status: degraded` means slower, not broken.","responses":{"200":{"description":"Instance can serve traffic","content":{"application/json":{"schema":{"description":"Instance can serve traffic","type":"object","properties":{"status":{"type":"string","enum":["ok","degraded"]},"redisCache":{"type":"string","enum":["up","down"]}}}}}}}}},"/products/{id}":{"get":{"summary":"Read a product (L1 → Redis → database)","tags":["products"],"description":"The `X-Cache` response header says which layer answered: `L1`, `L2`, or `MISS`.","parameters":[{"schema":{"type":"integer","minimum":1},"in":"path","name":"id","required":true,"description":"Cached in Redis as m01:product:v1:<id>"}],"responses":{"200":{"description":"Found","content":{"application/json":{"schema":{"description":"Found","type":"object","properties":{"id":{"type":"integer"},"name":{"type":"string"},"priceCents":{"type":"integer"},"updatedAt":{"type":"string","format":"date-time"}}}}}},"304":{"description":"Client copy is still current (If-None-Match matched the ETag)"},"404":{"description":"No product with this id","content":{"application/json":{"schema":{"description":"No product with this id","type":"object","properties":{"error":{"type":"string"}}}}}}}},"put":{"summary":"Update a product, then drop its cached copies","tags":["products"],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","minProperties":1,"additionalProperties":false,"properties":{"name":{"type":"string","minLength":1},"priceCents":{"type":"integer","minimum":0}}}}}},"parameters":[{"schema":{"type":"integer","minimum":1},"in":"path","name":"id","required":true,"description":"Cached in Redis as m01:product:v1:<id>"}],"responses":{"200":{"description":"Updated","content":{"application/json":{"schema":{"description":"Updated","type":"object","properties":{"id":{"type":"integer"},"name":{"type":"string"},"priceCents":{"type":"integer"},"updatedAt":{"type":"string","format":"date-time"}}}}}},"404":{"description":"No product with this id","content":{"application/json":{"schema":{"description":"No product with this id","type":"object","properties":{"error":{"type":"string"}}}}}}}}},"/stats":{"get":{"summary":"Cache counters for this process (reset on restart)","tags":["stats"],"responses":{"200":{"description":"Counters since the process started","content":{"application/json":{"schema":{"description":"Counters since the process started","type":"object","properties":{"l1Hits":{"type":"integer"},"l2Hits":{"type":"integer"},"misses":{"type":"integer"},"lookups":{"type":"integer"},"hitRate":{"type":"number","description":"(l1Hits + l2Hits) / lookups; null before the first lookup","nullable":true},"dbQueries":{"type":"integer","description":"Reads and writes that reached the database"},"l1Size":{"type":"integer","description":"Entries currently in this process’s L1"}}}}}}}}}},"servers":[{"url":"http://localhost:3000"}]}